+++ src/app/admin/event/[eventId]/_components/phases-card.tsx
@@ -24,12 +24,13 @@
 
 import { Card, CardContent } from '@/components/ui/card';
 import { Skeleton } from '@/components/ui/skeleton';
-import { Phase, PhaseType } from '@/generated/prisma/client';
-import { cn } from '@/lib/utils/cn';
 import {
   fetchPhasesForEvent,
   isPhasesSetupCompleted,
-} from '@/server/services/phase';
+} from '@/features/phase/services/phaseService';
+import { Phase, PhaseType } from '@/generated/prisma/client';
+import { cn } from '@/lib/utils/cn';
+
 import {
   BoltIcon,
   CheckIcon,
