+++ src/app/admin/event/[eventId]/page.tsx
@@ -1,7 +1,7 @@
 import { Button } from '@/components/ui/button';
 import { Card, CardContent } from '@/components/ui/card';
 import { Separator } from '@/components/ui/separator';
-import { getEventById } from '@/server/services/event';
+import { getEventById } from '@/features/event/services/eventService';
 import { PlusIcon } from '@heroicons/react/16/solid';
 import { PhasesCard } from './_components/phases-card';
