+++ src/app/api/auth/[...all]/route.ts
@@ -1,4 +1,4 @@
-import { auth } from '@/utils/auth';
+import { auth } from '@/lib/api/auth';
 import { toNextJsHandler } from 'better-auth/next-js';
 
 export const { POST, GET } = toNextJsHandler(auth);
