+++ src/app/company/layout.tsx
@@ -4,25 +4,36 @@ import {
   SidebarProvider,
   SidebarTrigger,
 } from '@/components/ui/sidebar';
+import { auth } from '@/lib/api/auth';
 import { CompanyRoutes } from '@/lib/consts/routes';
-import { HomeIcon } from '@heroicons/react/16/solid';
+import { headers } from 'next/headers';
+import { redirect } from 'next/navigation';
 
-export default function CompanyLayout({
+export default async function CompanyLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
+  const session = await auth.api.getSession({
+    headers: await headers(),
+  });
+
+  if (!session) {
+    redirect('/auth/login');
+  }
+
   return (
     <main>
       <SidebarProvider>
         <Sidebar
           items={[
             {
-              icon: HomeIcon,
+              icon: 'home',
               title: 'Dashboard',
               url: CompanyRoutes.DASHBOARD,
             },
           ]}
+          session={session}
           variant="inset"
         />
         <SidebarInset className="overflow-hidden">
